wallet/requirements.txt
Flask==2.0.1
pymongo==3.12.0
Flask-PyMongo==2.3.0
psycopg2-binary
